+++ module-bsp/board/rt1051/cellular/rt1051_cellular.hpp
@@ -39,7 +39,7 @@ namespace bsp {
 
         ssize_t Read(void *buf, size_t nbytes) override final;
 
-        uint32_t Write(void *buf, size_t nbytes) override final;
+        ssize_t Write(void *buf, size_t nbytes) override final;
